Adds test for creating landmark vote
