+++ spec/features/user/votes/create_spec.rb
@@ -8,24 +8,45 @@ describe 'A logged in user' do
     @recording = create(:recording, landmark: @landmark, user: @user)
   end
 
-  # it 'can add a vote to a specific landmarks recording' do
-  #   VCR.use_cassette('user_creates_vote', record: :new_episodes) do
-  #
-  #     visit landmark_path(@landmark.id)
-  #
-  #     within(".total-score") do
-  #       expect(page).to have_content("Rating: 0")
-  #     end
-  #
-  #     within(".up-vote") do
-  #       click_button "Up"
-  #     end
-  # 
-  #     expect(current_path).to eq(landmark_path(@landmark.id))
-  #
-  #     within(".total-score") do
-  #       expect(page).to have_content("Rating: 1")
-  #     end
-  #   end
-  # end
+  it 'can add a vote to a specific landmarks recording' do
+    VCR.use_cassette('user_creates_recording_vote', record: :new_episodes) do
+
+      visit landmark_path(@landmark.id)
+
+      within(".total-score") do
+        expect(page).to have_content("Rating: 0")
+      end
+
+      within(".up-vote") do
+        click_button "Up"
+      end
+
+      expect(current_path).to eq(landmark_path(@landmark.id))
+
+      within(".total-score") do
+        expect(page).to have_content("Rating: 1")
+      end
+    end
+  end
+
+  it 'can add a vote to a specific landmarks recording' do
+    VCR.use_cassette('user_creates_vote_landmark', record: :new_episodes) do
+
+      visit landmark_path(@landmark.id)
+
+      within(".landmark-total-score") do
+        expect(page).to have_content("Rating: 0")
+      end
+
+      within(".landmark-up-vote") do
+        click_button "Up"
+      end
+
+      expect(current_path).to eq(landmark_path(@landmark.id))
+
+      within(".landmark-total-score") do
+        expect(page).to have_content("Rating: 1")
+      end
+    end
+  end
 end
